+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
-## 5.0.14 (2022-10-03)
+## 5.0.15 (2022-10-05)
+### Feature
+* Added ipfabric.settings.SiteSeparation
+* Removed ipfabric.tools.UpdateSiteNames as this was removed in 4.4
+* Added ipfabric.tools.site_seperation_report.map_devices_to_rules
+* See [site_separation_tester.py](examples/tools/site_separation_tester.py)
+
+## 5.0.14 (2022-10-04)
 ### Feature
 * Added shared tools for parsing MAC addresses and return IP Fabric Format
 * Added examples for parsing MAC and Timestamps.